Mucked with the data type sizes and got them all cleaned up.  Still not sure about the ints all over the place for indexes and such, but the compiler give a warning for indexing with a char so I left them.  They don't seem to have much of a positive affect anyway.

+++ LampMatrix/OutputPins.cpp
@@ -8,7 +8,7 @@
 
 #include "Arduino.h"
 
-OutputPins::OutputPins(int pinCountIn, int pinsIn[])
+OutputPins::OutputPins(uint8_t pinCountIn, uint8_t pinsIn[])
 : pinNumber(&pinsIn[0]), pinCount(pinCountIn)
 {
 	pinValue = new int[pinCountIn];
@@ -60,7 +60,7 @@ void OutputPins::initializeDigitalPins(int value) {
     }
 }
 
-void OutputPins::setAnalogPin(int pinNdx, int value) {
+void OutputPins::setAnalogPin(uint8_t pinNdx, int value) {
     if (value == pinValue[pinNdx]) return;
 	analogWrite(pinNumber[pinNdx], value);
     pinValue[pinNdx] = value;
@@ -78,7 +78,7 @@ void OutputPins::setAnalogPins(int value) {
 	}
 }
 
-void OutputPins::setDigitalPin(int pinNdx, int value) {
+void OutputPins::setDigitalPin(uint8_t pinNdx, int value) {
     if (value == pinValue[pinNdx]) return;
     digitalWrite(pinNumber[pinNdx], value);
     pinValue[pinNdx] = value;
